fill documents and other, todo-accepted, saved

+++ src/main/java/com/terminalmock/test/repositories/enumsrepo/DocTypeRepo.java
@@ -3,5 +3,5 @@ package com.terminalmock.test.repositories.enumsrepo;
 import com.terminalmock.test.entities.enums.DocType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DocTypeRepo extends JpaRepository<DocType, Long> {
+public interface DocTypeRepo extends JpaRepository<DocType, Integer> {
 }
